// apps/api/src/modules/identity/domain/user/user-sort-field.enum.ts
export enum UserSortField {
  ID = 'id',
  FIRST_NAME = 'firstName',
  SECOND_NAME = 'secondName',
  LAST_NAME = 'lastName',
  FULL_NAME = 'fullName',
  EMAIL = 'email',
  STATUS = 'status',
  POSITION_ID = 'positionId',
  TEAM_ID = 'teamId',
  MANAGER_ID = 'managerId',
  CREATED_AT = 'createdAt',
  UPDATED_AT = 'updatedAt',
}